requirements.txt
# Backend Dependencies - Radar de Preço & CMV

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
supabase==2.9.0
sqlmodel==0.0.22

# Image Processing & OCR
pytesseract==0.3.13
Pillow==11.0.0
opencv-python==4.10.0.84

# Utilities
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.2

# News Scraping (Market Intel)
feedparser==6.0.11
beautifulsoup4==4.12.3

# Fuzzy Matching (for ingredient suggestions)
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.1
